[RENAME/#40] MainBottomNavigationBar -> OrbitBottomNavigationBar
