--- feature/navigator/src/main/java/com/yapp/navigator/NavItem.kt
@@ -1,40 +1,0 @@
-package com.yapp.navigator
-
-import androidx.annotation.DrawableRes
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
-
-@Composable
-fun RowScope.NavItem(
-    modifier: Modifier = Modifier,
-    selected: Boolean,
-    label: String,
-    @DrawableRes iconId: Int,
-    onClick: () -> Unit,
-) {
-    Column(
-        modifier = modifier
-            .weight(1f)
-            .fillMaxHeight()
-            .clickable(onClick = onClick),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        Icon(
-            painter = painterResource(id = iconId),
-            contentDescription = label,
-            tint = if (selected) Color.Blue else Color.Gray,
-        )
-        Text(text = label, color = if (selected) Color.Blue else Color.Black)
-    }
-}
